Modul h i cpp gotowy
