+++ Source.cpp
@@ -12,11 +12,10 @@ int main()
 			<< "\n1 - display_file_list"
 			<< "\n2 - display_harddrive"
 			<< "\n3 - display_harddrive_2"
-			<< "\n4 - create_empty_file"
-			<< "\n5 - delete_file"
-			<< "\n6 - load_file_from_Windows_and_save_on_harddrive"
-			<< "\n7 - create_file_based_on_string_system"
-			<< "\n8 - display_file"
+			<< "\n4 - delete_file"
+			<< "\n5 - load_file_from_Windows_and_save_on_harddrive"
+			<< "\n6 - create_file_based_on_string_system"
+			<< "\n7 - display_file"
 			<< "\nYour choice: ";
 		cin >> choice;
 
@@ -43,20 +42,15 @@ int main()
 		}
 		case 4:
 		{
-			create_empty_file(harddrive);
+			delete_file(harddrive);
 			break;
 		}
 		case 5:
 		{
-			delete_file(harddrive);
-			break;
-		}
-		case 6:
-		{
 			load_file_from_Windows_and_save_on_harddrive(harddrive);
 			break;
 		}
-		case 7:
+		case 6:
 		{
 			string filename = "Pliczek";
 			string type = "txt";
@@ -64,7 +58,7 @@ int main()
 			system_create_file_based_on_string(harddrive, filename, type, mystring);
 			break;
 		}
-		case 8:
+		case 7:
 		{
 			display_file(harddrive);
 			break;
@@ -72,6 +66,6 @@ int main()
 		default:
 			break;
 		}
-	} while (choice > 0 && choice <9);
+	} while (choice > 0 && choice <8);
 	return 0;
 }
